+++ .github/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <h3>New features since last release</h3>
 
+* Added validation for noise channel parameters. Invalid noise parameters now
+  raise a `ValueError`. [(#1357)](https://github.com/PennyLaneAI/pennylane/pull/1357)
+
 * PennyLane now has a `fourier` module, which hosts a [growing library 
   of methods](https://pennylane.readthedocs.io/en/stable/code/qml_fourier.html) 
   that help with investigating the Fourier representation of functions 
@@ -235,6 +238,9 @@ random_mat2 = rng.standard_normal(3, requires_grad=False)
 
 * Ising XX gate functionality added. [(#1194)](https://github.com/PennyLaneAI/pennylane/pull/1194)
 
+* Added Projector observable, which is available on all devices inheriting from the `QubitDevice` class.
+  [(#1356)](https://github.com/PennyLaneAI/pennylane/pull/1356)
+
 <h3>Improvements</h3>
 
 * The `benchmark` module was deleted, since it was outdated and is superseded by
@@ -358,6 +364,9 @@ random_mat2 = rng.standard_normal(3, requires_grad=False)
   [(#1332)](https://github.com/PennyLaneAI/pennylane/pull/1332)
 
 <h3>Documentation</h3>
+
+* Fix typo in the documentation of qml.templates.layers.StronglyEntanglingLayers 
+  [(#1367)](https://github.com/PennyLaneAI/pennylane/pull/1367)
 
 * Fixed typo on TensorFlow interface documentation [(#1312)](https://github.com/PennyLaneAI/pennylane/pull/1312)